+++ internal/statuspage/statuspagetypes/sort_test.go
@@ -1,4 +1,4 @@
-package statuspage
+package statuspagetypes
 
 import (
 	"reflect"
